Split channel forms into different files
